[Parley] refactor: Extract QuestPropertiesPopulator (#1228)
Moved quest field population logic (~100 lines) from PropertyPanelPopulator into QuestPropertiesPopulator.
Build succeeds, all 519 tests pass.
